Refactor extractor package
