+++ internal/extractor/openapi.go
@@ -7,21 +7,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"maps"
-	"slices"
 	"sort"
-	"strings"
 )
-
-// Schema is a single kind extracted from a Kubernetes OpenAPI v2 swagger
-// document. Schema is owned by the returned instance; the transforms mutate
-// it in place, so callers should treat it as single-use.
-type Schema struct {
-	Group   string
-	Version string
-	Kind    string
-	Schema  map[string]any
-}
 
 // ExtractOpenAPI walks a Kubernetes OpenAPI v2 swagger document and returns
 // one Schema per x-kubernetes-group-version-kind entry with all $refs inlined
@@ -75,10 +62,10 @@ func ExtractOpenAPI(data []byte) ([]Schema, []error) {
 				continue
 			}
 			out = append(out, Schema{
-				Group:   gvk.group,
-				Version: gvk.version,
-				Kind:    gvk.kind,
-				Schema:  schema,
+				Group:   gvk.Group,
+				Version: gvk.Version,
+				Kind:    gvk.Kind,
+				JSON:    schema,
 			})
 		}
 	}
@@ -96,23 +83,14 @@ func ExtractOpenAPI(data []byte) ([]Schema, []error) {
 	return out, errs
 }
 
-const (
-	jsonNullType  = "null"
-	jsonSchemaURI = "http://json-schema.org/schema#"
-)
+const jsonSchemaURI = "http://json-schema.org/schema#"
 
-type gvk struct {
-	group   string
-	version string
-	kind    string
-}
-
-func readGVKs(def map[string]any) []gvk {
+func readGVKs(def map[string]any) []GVK {
 	raw, ok := def["x-kubernetes-group-version-kind"].([]any)
 	if !ok {
 		return nil
 	}
-	var out []gvk
+	var out []GVK
 	for _, entry := range raw {
 		m, ok := entry.(map[string]any)
 		if !ok {
@@ -124,7 +102,7 @@ func readGVKs(def map[string]any) []gvk {
 		if k == "" || v == "" {
 			continue
 		}
-		out = append(out, gvk{group: g, version: v, kind: k})
+		out = append(out, GVK{Group: g, Version: v, Kind: k})
 	}
 	return out
 }
@@ -162,7 +140,7 @@ func buildSchema(name string, def map[string]any, defs map[string]any) (map[stri
 
 	// 6. additionalProperties:false including the root, so extra top-level
 	//    keys alongside apiVersion/kind/metadata/spec/... are rejected.
-	addAdditionalPropertiesFalse(inlined, false)
+	closeAdditionalProperties(inlined)
 
 	// 7. Strip remaining x-kubernetes-* extensions (keep preserve-unknown-fields).
 	stripVendorExtensions(inlined)
@@ -199,278 +177,5 @@ func deepCopyJSON(v any) (any, error) {
 		return out, nil
 	default:
 		return v, nil
-	}
-}
-
-// inlineRefs walks node and replaces every {$ref: "#/definitions/X"} object
-// with a deep-copy of defs[X]. Siblings of $ref follow the merge rules in
-// plans/extract-k8s.md: metadata siblings (description, default, example,
-// x-kubernetes-*) win over the inlined definition; structural siblings
-// (type, properties, items, etc.) are overridden by the inlined definition.
-// Cycles bail out to {type: object, x-kubernetes-preserve-unknown-fields: true}.
-func inlineRefs(node any, defs map[string]any, visiting map[string]bool) any {
-	switch n := node.(type) {
-	case map[string]any:
-		if refVal, ok := n["$ref"].(string); ok {
-			return resolveRef(n, refVal, defs, visiting)
-		}
-		for k, v := range n {
-			n[k] = inlineRefs(v, defs, visiting)
-		}
-		return n
-	case []any:
-		for i, v := range n {
-			n[i] = inlineRefs(v, defs, visiting)
-		}
-		return n
-	default:
-		return n
-	}
-}
-
-func resolveRef(node map[string]any, refVal string, defs map[string]any, visiting map[string]bool) any {
-	const prefix = "#/definitions/"
-	if !strings.HasPrefix(refVal, prefix) || len(refVal) == len(prefix) {
-		return overlaySiblings(unresolvedPlaceholder(refVal), node)
-	}
-	name := strings.TrimPrefix(refVal, prefix)
-
-	if visiting[name] {
-		return map[string]any{
-			"type":                                 "object",
-			"x-kubernetes-preserve-unknown-fields": true,
-		}
-	}
-
-	target, ok := defs[name]
-	if !ok {
-		return overlaySiblings(unresolvedPlaceholder(name), node)
-	}
-
-	clone, err := deepCopyJSON(target)
-	if err != nil {
-		return overlaySiblings(unresolvedPlaceholder(name), node)
-	}
-	next := copyVisiting(visiting)
-	next[name] = true
-	resolved := inlineRefs(clone, defs, next)
-
-	m, ok := resolved.(map[string]any)
-	if !ok {
-		return resolved
-	}
-	return overlaySiblings(m, node)
-}
-
-func unresolvedPlaceholder(name string) map[string]any {
-	return map[string]any{
-		"type":        "object",
-		"description": fmt.Sprintf("unresolved $ref: %s", name),
-	}
-}
-
-func copyVisiting(src map[string]bool) map[string]bool {
-	out := make(map[string]bool, len(src)+1)
-	maps.Copy(out, src)
-	return out
-}
-
-// overlaySiblings merges keys from the $ref host (other than $ref itself) onto
-// the inlined node. Structural keys from the inlined node always win; metadata
-// siblings (description, default, example, x-kubernetes-*, unknown keys) win
-// from the host side.
-func overlaySiblings(inlined map[string]any, host map[string]any) map[string]any {
-	for k, v := range host {
-		if k == "$ref" {
-			continue
-		}
-		if isStructuralKey(k) {
-			if _, has := inlined[k]; has {
-				continue
-			}
-		}
-		inlined[k] = v
-	}
-	return inlined
-}
-
-func isStructuralKey(k string) bool {
-	switch k {
-	case "type", "properties", "items", "required",
-		"oneOf", "allOf", "anyOf", "not",
-		"additionalProperties",
-		"enum", "format", "pattern",
-		"minimum", "maximum",
-		"minLength", "maxLength",
-		"minItems", "maxItems", "uniqueItems",
-		"minProperties", "maxProperties",
-		"multipleOf":
-		return true
-	}
-	return false
-}
-
-// injectGVK adds apiVersion and kind to the top-level properties if missing
-// and adds them to required. metadata is left as-is: absent kinds stay absent,
-// and inlined metadata is neither promoted to required nor demoted.
-func injectGVK(schema map[string]any) {
-	props, _ := schema["properties"].(map[string]any)
-	if props == nil {
-		props = map[string]any{}
-		schema["properties"] = props
-	}
-
-	required := toStringSlice(schema["required"])
-
-	if _, ok := props["apiVersion"]; !ok {
-		props["apiVersion"] = map[string]any{
-			"type":        "string",
-			"description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
-		}
-	}
-	if !containsString(required, "apiVersion") {
-		required = append(required, "apiVersion")
-	}
-
-	if _, ok := props["kind"]; !ok {
-		props["kind"] = map[string]any{
-			"type":        "string",
-			"description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
-		}
-	}
-	if !containsString(required, "kind") {
-		required = append(required, "kind")
-	}
-
-	schema["required"] = toAnySlice(required)
-}
-
-func toStringSlice(v any) []string {
-	arr, ok := v.([]any)
-	if !ok {
-		return nil
-	}
-	out := make([]string, 0, len(arr))
-	for _, item := range arr {
-		if s, ok := item.(string); ok {
-			out = append(out, s)
-		}
-	}
-	return out
-}
-
-func toAnySlice(in []string) []any {
-	out := make([]any, len(in))
-	for i, s := range in {
-		out[i] = s
-	}
-	return out
-}
-
-func containsString(haystack []string, needle string) bool {
-	return slices.Contains(haystack, needle)
-}
-
-// nullableOptional walks every object with properties and marks each property
-// that is not in required as nullable: scalar type becomes ["<t>", "null"],
-// list types gain "null" when missing, oneOf gets a {type: null} branch.
-// The transform is idempotent. Subtrees under x-kubernetes-preserve-unknown-fields
-// are skipped entirely.
-func nullableOptional(node any) {
-	switch n := node.(type) {
-	case map[string]any:
-		if preserve, _ := n["x-kubernetes-preserve-unknown-fields"].(bool); preserve {
-			return
-		}
-		props, _ := n["properties"].(map[string]any)
-		if props != nil {
-			required := toStringSlice(n["required"])
-			for name, prop := range props {
-				propMap, ok := prop.(map[string]any)
-				if !ok {
-					continue
-				}
-				if !containsString(required, name) {
-					markNullable(propMap)
-				}
-				// Recurse into the property's subtree so nested objects are
-				// processed with their own required array. Done here (not
-				// via the generic walk below) to avoid visiting the
-				// properties map twice.
-				nullableOptional(propMap)
-			}
-		}
-		for k, v := range n {
-			if k == "properties" {
-				continue
-			}
-			nullableOptional(v)
-		}
-	case []any:
-		for _, v := range n {
-			nullableOptional(v)
-		}
-	}
-}
-
-func markNullable(prop map[string]any) {
-	// type handling.
-	if t, ok := prop["type"]; ok {
-		switch tv := t.(type) {
-		case string:
-			if tv == jsonNullType {
-				return
-			}
-			prop["type"] = []any{tv, jsonNullType}
-			return
-		case []any:
-			for _, item := range tv {
-				if s, ok := item.(string); ok && s == jsonNullType {
-					return
-				}
-			}
-			prop["type"] = append(tv, jsonNullType)
-			return
-		}
-	}
-	// oneOf handling.
-	if oneOf, ok := prop["oneOf"].([]any); ok {
-		for _, branch := range oneOf {
-			bm, ok := branch.(map[string]any)
-			if !ok {
-				continue
-			}
-			if s, ok := bm["type"].(string); ok && s == jsonNullType {
-				return
-			}
-		}
-		prop["oneOf"] = append(oneOf, map[string]any{"type": jsonNullType})
-		return
-	}
-	// allOf/anyOf or otherwise unclassifiable: leave alone.
-}
-
-// stripVendorExtensions removes every x-kubernetes-* key from the tree except
-// x-kubernetes-preserve-unknown-fields, which carries structural meaning for
-// downstream validators.
-func stripVendorExtensions(node any) {
-	switch n := node.(type) {
-	case map[string]any:
-		for k := range n {
-			if !strings.HasPrefix(k, "x-kubernetes-") {
-				continue
-			}
-			if k == "x-kubernetes-preserve-unknown-fields" {
-				continue
-			}
-			delete(n, k)
-		}
-		for _, v := range n {
-			stripVendorExtensions(v)
-		}
-	case []any:
-		for _, v := range n {
-			stripVendorExtensions(v)
-		}
 	}
 }
